backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.2
django-channels==4.0.0
channels-redis==4.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
argon2-cffi==23.1.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
requests==2.31.0
python-dateutil==2.8.2
openpyxl==3.11.0
reportlab==4.0.8
